+++ src/providers/olark/olark-test.js
@@ -16,6 +16,10 @@
         email : 'zeus@segment.io'
     };
 
+
+    // Initialize
+    // ----------
+
     test('stores settings and adds olark.js on initialize', function () {
         expect(window.olark).not.to.exist;
 
@@ -25,6 +29,10 @@
         expect(window.olark).to.exist;
         expect(analytics.providers[0].settings.siteId).to.equal('x');
     });
+
+
+    // Identify
+    // --------
 
     test('updates visitor nickname on identify with the best name', function () {
         var spy = sinon.spy(window, 'olark');
@@ -62,12 +70,16 @@
         spy.restore();
     });
 
+
+    // Track
+    // -----
+
     test('logs event to operator on track if `track` setting is true', function () {
         analytics.providers[0].settings.track = true;
         var spy = sinon.spy(window, 'olark');
         analytics.track(event, properties);
         expect(spy).to.have.been.calledWith('api.chat.sendNotificationToOperator', sinon.match({
-            body : 'Visitor triggered "'+event+'".'
+            body : 'visitor triggered "'+event+'"'
         }));
 
         spy.restore();
@@ -77,6 +89,30 @@
         analytics.providers[0].settings.track = false;
         var spy = sinon.spy(window, 'olark');
         analytics.track(event, properties);
+        expect(spy).not.to.have.been.called;
+
+        spy.restore();
+    });
+
+
+    // Pageview
+    // --------
+
+    test('logs event to operator on pageview if `pageview` setting is true', function () {
+        analytics.providers[0].settings.pageview = true;
+        var spy = sinon.spy(window, 'olark');
+        analytics.pageview();
+        expect(spy).to.have.been.calledWith('api.chat.sendNotificationToOperator', sinon.match({
+            body : 'looking at ' + window.location.href
+        }));
+
+        spy.restore();
+    });
+
+    test('doesnt log event to operator on pageview if `pageview` setting is false', function () {
+        analytics.providers[0].settings.pageview = false;
+        var spy = sinon.spy(window, 'olark');
+        analytics.pageview();
         expect(spy).not.to.have.been.called;
 
         spy.restore();